requirements.txt
gspread
feedparser
pandas
google-auth